finlit-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
scikit-learn